Extract portfolio allocation panel state
