+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -34,3 +34,4 @@
 | LWB-R030 | Portfolio record evidence rail identity duplication cleanup | Hardened | Dedicated record screens already show portfolio identity in the page header and left navigation, but the evidence rail repeated the raw portfolio id as a fact and embedded it inside source-coverage detail text. This added technical noise, made the rail less dense, and repeated context without helping the advisor decide whether the record screen was ready to use. | Removed the duplicate portfolio-id fact from the record evidence view model, renamed the remaining rail facts to front-office language (`Currency`, `Review Area`), and rewrote positions/transaction source details to describe available review coverage without repeating the portfolio id. Kept portfolio ids only in workflow links where routing requires them. | Continue reducing portfolio record-screen rail density by moving lower-priority governance facts into a drawer only if canonical visual review still shows side-rail overload. | `src/apps/portfolio/portfolio-record-evidence-view-model.ts`, `tests/unit/portfolio-record-evidence-view-model.test.ts`, `tests/unit/portfolio-record-evidence-rail.test.tsx`; validation via `npm run test -- tests/unit/portfolio-record-evidence-view-model.test.ts tests/unit/portfolio-record-evidence-rail.test.tsx`, `npm run typecheck`, `git diff --check` |
 | LWB-R031 | Portfolio record rail workflow and source-language tightening | Hardened | The record evidence rail still exposed service-shaped wording such as `Core positions inventory` and `Core transaction ledger`, and adjacent-workflow links could point back to the current record screen. This weakened front-office scan value and kept low-value navigation chrome in the rail. | Reframed record evidence sources with advisor-readable book and activity language, added `Positions` to adjacent workflows where useful, and filtered the active record screen out of the adjacent-workflow list so the rail only offers real next navigation choices. Updated view-model and component tests to guard the current-screen filtering and source-copy cleanup. | Continue removing source/service labels from visible portfolio copy where the source posture can be expressed as book, holdings, transaction, activity, reporting, or mandate evidence without losing truth. | `src/apps/portfolio/portfolio-record-evidence-view-model.ts`, `tests/unit/portfolio-record-evidence-view-model.test.ts`, `tests/unit/portfolio-record-evidence-rail.test.tsx`; validation via `npm run test -- tests/unit/portfolio-record-evidence-view-model.test.ts tests/unit/portfolio-record-evidence-rail.test.tsx`, `npm run typecheck`, `git diff --check` |
 | LWB-R032 | Portfolio workspace client recovery test race | Hardened | The full Feature Lane exposed an async race in the portfolio workspace recovery test: the test could assert the recovered shell render, finish, reset API mocks, and only then let the summary-details effect run. That made the suite order-sensitive and produced an unhandled `finally` failure unrelated to product behavior. | Updated the recovery test to wait for the summary-details Gateway call before asserting its payload, keeping the test aligned with the component's shell-then-details loading sequence. | Continue preferring effect-aware waits for client data-loading tests so mock cleanup cannot race pending Workbench requests. | `tests/unit/portfolio-workspace-client.test.tsx`; validation via `npm run test -- tests/unit/portfolio-workspace-client.test.tsx tests/unit/portfolio-record-evidence-view-model.test.ts tests/unit/portfolio-record-evidence-rail.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check` |
+| LWB-R033 | Portfolio allocation panel state extraction | Hardened | `portfolio-allocation-panel.tsx` still mixed allocation rendering with Gateway look-through probing, response caching, active-dimension repair, chart-mode state, hover state, and selection cleanup. The visible copy also exposed implementation-shaped "pending source support" language for unavailable dimensions. | Extracted allocation panel orchestration into `use-portfolio-allocation-panel-state.ts`, leaving the panel focused on rendering controls and visuals. Replaced unavailable-dimension copy with allocation-coverage language and added component coverage for the front-office label and no-regression posture. | Continue splitting large portfolio grids by moving row shaping, sorting, and drawer payload construction into view-model modules before changing visible table behavior. | `src/apps/portfolio/components/portfolio-allocation-panel.tsx`, `src/apps/portfolio/components/use-portfolio-allocation-panel-state.ts`, `tests/unit/portfolio-allocation-panel.test.tsx`; validation via `npm run test -- tests/unit/portfolio-allocation-panel.test.tsx tests/unit/portfolio-allocation-view-model.test.ts`, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build` |
